# packages/backend/src/openapi-merchandise.yaml
openapi: 3.0.0
info:
  title: PSE 2025 Merchandise Shop API
  version: 1.0.0
  description: API documentation for PSE 2025 merchandise shop, shopping cart, and orders
  contact:
    name: PSE Motors
    email: api@psemotors.com

servers:
  - url: http://localhost:4000
    description: Development server

tags:
  - name: Products
    description: Merchandise product endpoints
  - name: Cart
    description: Shopping cart endpoints
  - name: Orders
    description: Order management endpoints

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          example: cmqkp1fzx0001rx9kla8mnf47
        name:
          type: string
          example: PSE Motors T-Shirt
        description:
          type: string
          example: Premium cotton t-shirt with PSE Motors logo
        price:
          type: string
          example: "29.99"
        category:
          type: string
          enum: [apparel, accessories, collectibles]
          example: apparel
        imageUrl:
          type: string
          example: /images/products/tshirt-black.jpg
        stock:
          type: integer
          example: 50
        available:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CartItem:
      type: object
      properties:
        id:
          type: string
        cartId:
          type: string
        productId:
          type: string
        quantity:
          type: integer
          example: 2
        createdAt:
          type: string
          format: date-time
        product:
          $ref: '#/components/schemas/Product'

    Cart:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
          format: uuid
        userId:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CartWithTotals:
      allOf:
        - $ref: '#/components/schemas/Cart'
        - type: object
          properties:
            itemCount:
              type: integer
              description: Total number of items in cart
              example: 3
            subtotal:
              type: string
              description: Subtotal before tax
              example: "99.99"
            tax:
              type: string
              description: Tax amount (10%)
              example: "10.00"
            total:
              type: string
              description: Total amount including tax
              example: "109.99"

    OrderItem:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        productId:
          type: string
        quantity:
          type: integer
        price:
          type: string
          description: Price at time of order
        createdAt:
          type: string
          format: date-time
        product:
          $ref: '#/components/schemas/Product'

    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
          example: ORD-1234567890
        sessionId:
          type: string
          format: uuid
        totalAmount:
          type: string
          example: "109.99"
        status:
          type: string
          enum: [pending, completed, cancelled]
          example: pending
        customerEmail:
          type: string
          format: email
          nullable: true
          example: customer@example.com
        customerName:
          type: string
          nullable: true
          example: John Doe
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 45
        totalPages:
          type: integer
          example: 3

    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation error
        message:
          type: string
          example: Invalid request parameters

paths:
  /api/products:
    get:
      summary: List all products
      description: Get a paginated list of all merchandise products with optional filters
      tags:
        - Products
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [apparel, accessories, collectibles]
          description: Filter by product category
        - in: query
          name: minPrice
          schema:
            type: number
          description: Minimum price filter
        - in: query
          name: maxPrice
          schema:
            type: number
          description: Maximum price filter
        - in: query
          name: available
          schema:
            type: boolean
          description: Filter by availability
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Items per page
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/products/{id}:
    get:
      summary: Get product details
      description: Get detailed information about a specific product
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/cart/{sessionId}:
    get:
      summary: Get shopping cart
      description: Get the shopping cart for a session with items and totals
      tags:
        - Cart
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID
      responses:
        '200':
          description: Shopping cart with items and totals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartWithTotals'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Add item to cart
      description: Add a product to the shopping cart
      tags:
        - Cart
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - quantity
              properties:
                productId:
                  type: string
                  description: Product ID to add
                quantity:
                  type: integer
                  minimum: 1
                  description: Quantity to add
                  example: 2
      responses:
        '200':
          description: Item added to cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartWithTotals'
        '400':
          description: Invalid request or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Clear cart
      description: Remove all items from the shopping cart
      tags:
        - Cart
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID
      responses:
        '200':
          description: Cart cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Cart cleared successfully
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/cart/{sessionId}/{itemId}:
    put:
      summary: Update cart item quantity
      description: Update the quantity of an item in the cart
      tags:
        - Cart
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID
        - in: path
          name: itemId
          required: true
          schema:
            type: string
          description: Cart item ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 1
                  description: New quantity
                  example: 3
      responses:
        '200':
          description: Cart item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartWithTotals'
        '400':
          description: Invalid request or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Remove item from cart
      description: Remove a specific item from the shopping cart
      tags:
        - Cart
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID
        - in: path
          name: itemId
          required: true
          schema:
            type: string
          description: Cart item ID
      responses:
        '200':
          description: Item removed from cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartWithTotals'
        '404':
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/cart/{sessionId}/checkout:
    post:
      summary: Checkout and create order
      description: Create an order from the current cart contents
      tags:
        - Cart
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                customerEmail:
                  type: string
                  format: email
                  description: Customer email address
                  example: customer@example.com
                customerName:
                  type: string
                  description: Customer name
                  example: John Doe
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Cart is empty or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/orders/{orderNumber}:
    get:
      summary: Get order details
      description: Get detailed information about a specific order
      tags:
        - Orders
      parameters:
        - in: path
          name: orderNumber
          required: true
          schema:
            type: string
          description: Order number
          example: ORD-1234567890
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'